early_access/a2f-3d-authoring-sample-app/requirements.txt
grpcio-tools
protobuf==5.28.1
pydantic
grpcio
numpy
scipy
matplotlib
